requirements.txt
django
python-dotenv
djangorestframework
psycopg2
cryptography
django-redis